kc-sactionlist-api-feature_web_scrapping/scrapper/requirements.txt
selenium==4.5.0
PyYAML==6.0
webdriver-manager==3.8.3
bs4
html5lib
psycopg2
PyPDF2
datetime
pandas
mysql-connector-python
re
